+++ index.html
@@ -241,8 +241,8 @@
       </div>
     </section>
   </template>
-  <script src="js/main.js"></script>
-  <script src="js/functions.js"></script>
+  <script type="module" src="./js/main.js"></script>
+  <script src="./js/functions.js"></script>
 </body>
 
 </html>
